Now you can pass the MAX_STICKINESS as a parameter upon execution. Also added two to the power of the backoff_stage as the number of frames successfuly transmitted.

+++ defs/display.h
@@ -9,7 +9,7 @@
 #include <fstream>
 using namespace std;
 
-void display(int sta, int tx, int collisions){
+void display(int sta, long int tx,  long int collisions){
 	ofstream output;
 	output.open("results.txt", ios::app);
 	output << sta << " " << tx << " " << collisions << endl;
